add novels

+++ train.py
@@ -112,7 +112,7 @@ for step in range(max_steps):
     for i in range(grad_accum_steps):
         x, y = train_loader.next_batch()
         x,y = x.to(device), y.to(device)
-        with torch.autocast(device_type=device, dtype=torch.float16):
+        with torch.autocast(device_type=device, dtype=torch.bfloat16):
             logits, loss = model(x,y)
         loss = loss / grad_accum_steps
         loss_accum += loss.detach()
